# estrado-pjud-service/openapi.json
{
  "components": {
    "schemas": {
...
      "FamiliaCaseFilter": {
        "properties": {
          "rit": {
            "title": "Rit",
            "type": "string"
          },
          "year": {
            "title": "Year",
            "type": "string"
          }
        },
        "required": [
          "rit",
          "year"
        ],
        "title": "FamiliaCaseFilter",
        "type": "object"
      },
      "FamiliaCaso": {
        "properties": {
          "caratulado": {
            "title": "Caratulado",
            "type": "string"
          },
          "estado": {
            "title": "Estado",
            "type": "string"
          },
          "fecha_ingreso": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Ingreso"
          },
          "materia": {
            "title": "Materia",
            "type": "string"
          },
          "rit": {
            "title": "Rit",
            "type": "string"
          },
          "tribunal": {
            "title": "Tribunal",
            "type": "string"
          }
        },
        "required": [
          "rit",
          "tribunal",
          "caratulado",
          "materia",
          "estado"
        ],
        "title": "FamiliaCaso",
        "type": "object"
      },
      "FamiliaSyncRequest": {
        "properties": {
          "auth_type": {
            "default": "clave_pj",
            "enum": [
              "clave_pj",
              "clave_unica"
            ],
            "title": "Auth Type",
            "type": "string"
          },
          "cases": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FamiliaCaseFilter"
            },
            "maxItems": 10,
            "title": "Cases",
            "type": "array"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "rut": {
            "title": "Rut",
            "type": "string"
          }
        },
        "required": [
          "rut",
          "password"
        ],
        "title": "FamiliaSyncRequest",
        "type": "object"
      },
      "FamiliaSyncResponse": {
        "properties": {
          "casos": {
            "items": {
              "$ref": "#/components/schemas/FamiliaCaso"
            },
            "title": "Casos",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_code": {
            "anyOf": [
              {
                "enum": [
                  "invalid_credentials",
                  "session_error",
                  "no_cases",
                  "parse_error",
                  "blocked"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "casos"
        ],
        "title": "FamiliaSyncResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PrivateCauseResolution": {
        "additionalProperties": false,
        "properties": {
          "case_number": {
            "title": "Case Number",
            "type": "string"
          },
          "case_type": {
            "const": "rit",
            "default": "rit",
            "title": "Case Type",
            "type": "string"
          },
          "detail": {
            "$ref": "#/components/schemas/SanitizedCaseDetail"
          },
          "matter": {
            "const": "familia",
            "default": "familia",
            "title": "Matter",
            "type": "string"
          },
          "movements": {
            "items": {
              "$ref": "#/components/schemas/SanitizedMovement"
            },
            "title": "Movements",
            "type": "array"
          },
          "tribunal_code": {
            "exclusiveMinimum": 0.0,
            "title": "Tribunal Code",
            "type": "integer"
          },
          "tribunal_label": {
            "title": "Tribunal Label",
            "type": "string"
          }
        },
        "required": [
          "case_number",
          "tribunal_code",
          "tribunal_label",
          "detail",
          "movements"
        ],
        "title": "PrivateCauseResolution",
        "type": "object"
      },
      "PrivateCauseResolutionRequest": {
        "additionalProperties": false,
        "description": "Closed request boundary for one explicitly selected Familia candidate.",
        "properties": {
          "auth_type": {
            "const": "clave_pj",
            "default": "clave_pj",
            "title": "Auth Type",
            "type": "string"
          },
          "case_number": {
            "maxLength": 128,
            "pattern": "^[A-Z]+-[0-9]+-[0-9]{4}$",
            "title": "Case Number",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "rut": {
            "title": "Rut",
            "type": "string"
          },
          "tribunal_code": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tribunal Code"
          },
          "tribunal_label": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Tribunal Label",
            "type": "string"
          }
        },
        "required": [
          "rut",
          "password",
          "case_number",
          "tribunal_label"
        ],
        "title": "PrivateCauseResolutionRequest",
        "type": "object"
      },
      "PrivateCauseResolutionResult": {
        "additionalProperties": false,
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "enum": [
                  "private_not_found",
                  "private_ambiguous",
                  "private_identifier_mismatch",
                  "private_tribunal_mismatch",
                  "private_evidence_incomplete",
                  "credential_invalid",
                  "session_expired",
                  "waf",
                  "timeout",
                  "upstream_changed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "resolution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrivateCauseResolution"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "ok"
        ],
        "title": "PrivateCauseResolutionResult",
        "type": "object"
      },
      "SanitizedCaseDetail": {
        "additionalProperties": false,
        "properties": {
          "caption": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Caption",
            "type": "string"
          },
          "filed_at": {
            "anyOf": [
              {
                "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filed At"
          },
          "status": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Status",
            "type": "string"
          },
          "subject": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          }
        },
        "required": [
          "caption",
          "status"
        ],
        "title": "SanitizedCaseDetail",
        "type": "object"
      },
      "SanitizedMovement": {
        "additionalProperties": false,
        "properties": {
          "date": {
            "anyOf": [
              {
                "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "kind": {
            "const": "status",
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "label"
        ],
        "title": "SanitizedMovement",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "estrado-pjud-service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/familia/resolve-private": {
      "post": {
        "operationId": "familia_resolve_private_api_v1_familia_resolve_private_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivateCauseResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateCauseResolutionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Familia Resolve Private",
        "tags": [
          "familia"
        ]
      }
    },
    "/api/v1/familia/sync": {
      "post": {
        "operationId": "familia_sync_api_v1_familia_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamiliaSyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamiliaSyncResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Familia Sync",
        "tags": [
          "familia"
        ]
      }
    },
...
  }
}